requirements.txt
crawl4ai
python-dotenv
pydantic
google-generativeai>=0.4.0